# (fix): #1038 long db build when notes contain URLs with https://