--- a/org-roam.el
+++ b/org-roam.el
@@ -532,8 +532,7 @@ PATH should be the root from which to compute the relativity."
       ;; Loop over links
       (while (re-search-forward org-roam--org-link-bracket-typed-re (point-max) t)
         (setq link (match-string 2))
-        (when (and (f-relative-p link)
-                   (file-attributes link))
+        (when (f-relative-p link)
           (save-excursion
             (goto-char (match-beginning 2))
             (delete-region (match-beginning 2)