###### Motivation for this change
See the detail in issue #1038 

With using [test files](https://github.com/nobiot/org-roam-performance-test-v1.2.1-Windows) of five (5) `.org` notes with `https` URL links, the following performance improvement is observed:

---

### Steps / code executed for benchmark

```
#+begin_src emacs-lisp
;;(org-roam-db--close-all)
;; delete the db file manually
(benchmark 1 '(org-roam-db-build-cache))
#+end_src
```

### Results
[2020-09-15 Tue]
Before change
Elapsed time: 123.473523s (0.506992s in 12 GCs)

After change
1st: Elapsed time: 0.253582s
2nd: Elapsed time: 0.240944s
3rd: Elapsed time: 0.568807s (0.043532s in 1 GCs)